+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Add minimal CLI with version and help flags
 - Add the `ignore-re-iface` option which can be used (multiple times) to ignore network interfaces based on regular expression pattern matching
 
+### Fixed
+
+- Fix issues with wrongly resolved hostnames from misinterpreted PTR records.
 
 ## [0.4.1]