+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Add minimal CLI with version and help flags
 - Add the `ignore-re-iface` option which can be used (multiple times) to ignore network interfaces based on regular expression pattern matching
 
+### Changed
+
+- Decreasing verbosity is now done with `g` instead of `c`
+
 ### Fixed
 
 - Fix issues with wrongly resolved hostnames from misinterpreted PTR records.